requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
idna==3.4
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
waitress==2.1.0
gunicorn==19.7.1